frontend: Add a settings button to the root cluster chooser dialog
So the user can access the settings even if a cluster is not yet
configured.
